+++ Test/Input/Bundler/BundleBug4/out/cb-bundle.js
@@ -1,6 +1,6 @@
 (function(undefined) {
     "use strict";
-    var __export_ns;
+    var __export_ns = {};
     function fun_lib2() {
         console.log("fun");
     }
@@ -9,7 +9,7 @@
             fun_lib2();
         }
         ns.fun = fun;
-    })(__export_ns || (__export_ns = {}));
+    })(__export_ns);
     __export_ns.fun();
 }).call(this);