+++ Test/Input/Bundler/EsmBootstrap/esm.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Carousel = void 0;
+var Carousel = /** @class */ (function () {
+  function Carousel() {
+    console.log("Ok");
+  }
+  return Carousel;
+})();
+exports.Carousel = Carousel;