+++ Test/Input/Bundler/EsmBootstrap/out/cb-bundle.js
@@ -1,0 +1,12 @@
+(function(undefined) {
+    "use strict";
+    var Carousel_esm;
+    Carousel_esm = function() {
+        function Carousel() {
+            console.log("Ok");
+        }
+        return Carousel;
+    }();
+    new Carousel_esm();
+}).call(this);
+