+++ Test/Input/Bundler/EsmBootstrap/out/cbm-bundle.js
@@ -1,0 +1,12 @@
+(function(e) {
+    "use strict";
+    var s;
+    s = function() {
+        function e() {
+            console.log("Ok");
+        }
+        return e;
+    }();
+    new s();
+}).call(this);
+