+++ Test/Input/Bundler/EsmBootstrap/out/cm-bundle.js
@@ -1,0 +1,1 @@
+(function(e){"use strict";var s;s=function(){function e(){console.log("Ok")}return e}();new s}).call(this)