+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Optimize TypeScript enums
 - dead code elimination of result unused `new Map()`
 
+### Fixed
+
+- Bundler with import default when there is not default export
+
 ## 0.4.0
 
 ### Added