+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+## 0.5.0
+
 ### Added
 
 - Optimize TypeScript enums